fix: added time for hiring

+++ services/jobs/data/job.go
@@ -16,7 +16,7 @@ type Job struct {
 	LikedBy     []primitive.ObjectID `json:"likedBy" bson:"likedBy"`
 	AppliedBy   []primitive.ObjectID `json:"appliedBy" bson:"appliedBy"`
 	Hired       *primitive.ObjectID  `json:"hired" bson:"hired"`
-	HiredAt     int32                `json:"hiredAt" bson:"hiredAt"`
+	HiredAt     int64                `json:"hiredAt" bson:"hiredAt"`
 	Budget      float32              `json:"budget" bson:"budget" validate:"required"`
 	BudgetMeta  string               `json:"budgetMeta" bson:"budgetMeta" validate:"required"`
 	Username    string               `json:"username" bson:"username" validate:"required"`
